Clear file selection after upload.
This is especially for Chrome which will complain and get stuck if a previously selected file was altered on disk. Re-selecting the same file didn't help, you had to open the dialog, cancel or select something else to clear the selection, only then you could re-select the file and upload it.

Note: file upload button is entirely broken in Safari 14.0.1. This is definitely a Safari bug. Use drag-and-drop or just use another browser altogether.